fix: UI와 로직 분리
