+++ src/app/(common)/_home/CardList.tsx
@@ -2,75 +2,38 @@
 
 import CreateGatheringsModal from "@/app/(common)/_home/_components/CreateGatheringsModal";
 import GatheringFilters from "@/app/(common)/_home/_components/GatheringFilters";
-import { useCheckAuth } from "@/app/(common)/_home/_hooks/useCheckAuth";
 import Button from "@/components/Button";
 import Card from "@/components/Card";
 import { LoginPopup } from "@/components/Popup";
 import ServiceTab from "@/components/ServiceTab";
 import { useInfiniteScroll } from "@/hooks/useInfiniteScroll";
+import { useFilters } from "@/app/(common)/_home/_hooks/useFilters";
 import { useGatherings } from "@/app/(common)/_home/_hooks/useGathering";
+import { useAuth } from "@/app/(common)/_home/_hooks/useAuth";
 import GatheringLogo from "@/images/gathering_logo.svg";
-import { useUserStore } from "@/stores/useUserStore";
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useState, useEffect } from "react";
 
 export default function CardList() {
-  const searchParams = useSearchParams();
-  const router = useRouter();
-  const pathname = usePathname();
-  const searchParamsString = searchParams.toString();
-
-  const filters = useMemo(
-    () => ({
-      location: searchParams.get("location") || undefined,
-      date: searchParams.get("date") || undefined,
-      sortBy: searchParams.get("sortBy") || undefined,
-      type: searchParams.get("type") || "DALLAEMFIT",
-    }),
-    [searchParamsString],
-  );
-
+  const { filters, handleFilterChange } = useFilters(); // 필터 관리 적용
+  const { filteredCards, fetchNextPage, hasNextPage, isFetchingNextPage, isLoading, dataFetched } =
+    useGatherings(filters);
+  const { checkAuth, isAuthModalOpen, setAuthModalOpen, shouldOpenCreateModal, setShouldOpenCreateModal } = useAuth(); // 로그인 체크 적용
   const [isModalOpen, setIsModalOpen] = useState(false);
-  const { filteredCards, fetchNextPage, hasNextPage, isFetchingNextPage, isLoading } = useGatherings(filters);
-
-  // 로그인 체크 훅
-  const { checkAuth, isAuthModalOpen, setAuthModalOpen } = useCheckAuth();
-  const shouldOpenCreateModal = useUserStore((state) => state.shouldOpenCreateModal);
-  const setShouldOpenCreateModal = useUserStore((state) => state.setShouldOpenCreateModal);
 
   useEffect(() => {
     if (shouldOpenCreateModal) {
       setIsModalOpen(true);
-      setShouldOpenCreateModal(false); // 상태 초기화
+      setShouldOpenCreateModal(false); // 로그인 후 다시 열리지 않도록 초기화
     }
-  }, [shouldOpenCreateModal]);
+  }, [shouldOpenCreateModal, setShouldOpenCreateModal]);
 
   const handleOpen = () => {
-    checkAuth(() => setIsModalOpen(true)); // 로그인 여부 확인 후 실행
+    checkAuth(() => {
+      setIsModalOpen(true);
+      setShouldOpenCreateModal(false);
+    });
   };
   const handleClose = () => setIsModalOpen(false);
-
-  // 필터를 업데이트 : searchParamsString을 사용하여 불필요한 재생성 방지
-  const handleFilterChange = useCallback(
-    (newFilter: Partial<typeof filters>) => {
-      const params = new URLSearchParams(searchParams.toString());
-
-      Object.entries(newFilter).forEach(([key, value]) => {
-        if (value) {
-          params.set(key, value);
-        } else {
-          params.delete(key);
-        }
-      });
-
-      const newParamsString = params.toString();
-
-      if (newParamsString !== searchParamsString) {
-        router.push(`${pathname}?${newParamsString}`);
-      }
-    },
-    [searchParamsString, router, pathname],
-  );
 
   // 무한 스크롤 훅 사용
   const { observerRef } = useInfiniteScroll({ fetchNextPage, hasNextPage, isFetchingNextPage });
@@ -87,7 +50,7 @@ export default function CardList() {
       <div className="relative mt-6">
         <div className="flex flex-row">
           <ServiceTab
-            searchParams={searchParams}
+            searchParams={new URLSearchParams(filters.toString())}
             onCategoryChange={(type) => {
               handleFilterChange({ type });
             }}
