+++ src/app/(common)/_home/_hooks/useFilters.ts
@@ -1,0 +1,43 @@
+import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { useMemo, useCallback } from "react";
+
+export const useFilters = () => {
+  const searchParams = useSearchParams();
+  const router = useRouter();
+  const pathname = usePathname();
+  const searchParamsString = searchParams.toString();
+
+  // searchParams에서 필터 값을 추출
+  const filters = useMemo(() => {
+    return {
+      location: searchParams.get("location") || undefined,
+      date: searchParams.get("date") || undefined,
+      sortBy: searchParams.get("sortBy") || undefined,
+      type: searchParams.get("type") || "DALLAEMFIT",
+    };
+  }, [searchParamsString]); // `searchParamsString` 대신 `searchParams` 사용
+
+  // 필터를 업데이트 : searchParamsString을 사용하여 불필요한 재생성 방지
+  const handleFilterChange = useCallback(
+    (newFilter: Partial<typeof filters>) => {
+      const params = new URLSearchParams(searchParams.toString());
+
+      Object.entries(newFilter).forEach(([key, value]) => {
+        if (value) {
+          params.set(key, value);
+        } else {
+          params.delete(key);
+        }
+      });
+
+      const newParamsString = params.toString();
+
+      if (newParamsString !== searchParamsString) {
+        router.push(`${pathname}?${newParamsString}`);
+      }
+    },
+    [searchParamsString, router, pathname],
+  );
+
+  return { filters, handleFilterChange };
+};
